+++ FuturesTests/FutureTests.swift
@@ -11,10 +11,193 @@ internal enum TestError: ErrorType {
 struct FutureSharedExamplesContext {
   static let Future = "future"
   static let Promise = "promise"
+  static let Value = "value"
 }
 
 class FutureSharedExamplesConfiguration: QuickConfiguration {
   override class func configure(configuration: Configuration) {
+    sharedExamples("failure case") { (sharedExampleContext: SharedExampleContext) in
+      var request: Future<String>!
+      var successSentinels: [String?]!
+      var failureSentinels: [ErrorType?]!
+      var cancelSentinels: [Bool?]!
+      
+      beforeEach {
+        request = sharedExampleContext()[FutureSharedExamplesContext.Future] as? Future<String>
+        
+        cancelSentinels = [nil, nil, nil]
+        successSentinels = [nil, nil, nil]
+        failureSentinels = [nil, nil, nil]
+      }
+      
+      context("when calling onSuccess") {
+        beforeEach {
+          for idx in 0..<successSentinels.count {
+            request.onSuccess { result in
+              successSentinels[idx] = result
+            }
+          }
+        }
+        
+        it("should not call the closures") {
+          expect(successSentinels.filter({ $0 == nil }).count).to(equal(successSentinels.count))
+        }
+      }
+      
+      context("when calling onFailure") {
+        beforeEach {
+          for idx in 0..<successSentinels.count {
+            request.onFailure { error in
+              failureSentinels[idx] = error
+            }
+          }
+        }
+        
+        it("should immediately call the closures") {
+          expect(failureSentinels.filter({ $0 != nil }).count).to(equal(failureSentinels.count))
+        }
+        
+        it("should pass the right error") {
+          for fSentinel in failureSentinels {
+            expect(fSentinel as? TestError).to(equal(TestError.SimpleError))
+          }
+        }
+      }
+      
+      context("when calling onCancel") {
+        beforeEach {
+          for idx in 0..<cancelSentinels.count {
+            request.onCancel {
+              cancelSentinels[idx] = true
+            }
+          }
+        }
+        
+        it("should not call the closures") {
+          expect(cancelSentinels.filter({ $0 == nil}).count).to(equal(cancelSentinels.count))
+        }
+      }
+      
+      context("when calling onCompletion") {
+        beforeEach {
+          for idx in 0..<successSentinels.count {
+            request.onCompletion { result in
+              switch result {
+              case .Success(let value):
+                successSentinels[idx] = value
+              case .Error(let error):
+                failureSentinels[idx] = error
+              default:
+                break
+              }
+            }
+          }
+        }
+        
+        it("should immediately call the closures") {
+          expect(failureSentinels.filter({ $0 != nil }).count).to(equal(failureSentinels.count))
+        }
+        
+        it("should pass the right error") {
+          for fSentinel in failureSentinels {
+            expect(fSentinel as? TestError).to(equal(TestError.SimpleError))
+          }
+        }
+        
+        it("should not pass a value") {
+          expect(successSentinels.filter({ $0 == nil }).count).to(equal(successSentinels.count))
+        }
+      }
+    }
+    
+    sharedExamples("success case") { (sharedExampleContext: SharedExampleContext) in
+      var request: Future<String>!
+      var value: String!
+      var successSentinels: [String?]!
+      var failureSentinels: [ErrorType?]!
+      var cancelSentinels: [Bool?]!
+      
+      beforeEach {
+        request = sharedExampleContext()[FutureSharedExamplesContext.Future] as? Future<String>
+        value = sharedExampleContext()[FutureSharedExamplesContext.Value] as? String
+        
+        successSentinels = [nil, nil, nil]
+        failureSentinels = [nil, nil, nil]
+        cancelSentinels = [nil, nil, nil]
+      }
+      
+      context("when calling onSuccess") {
+        beforeEach {
+          for idx in 0..<successSentinels.count {
+            request.onSuccess({ result in
+              successSentinels[idx] = result
+            })
+          }
+        }
+        
+        it("should immediately call the closures") {
+          expect(successSentinels.filter({ $0 != nil }).count).to(equal(successSentinels.count))
+        }
+        
+        it("should pass the right value") {
+          expect(successSentinels).to(allPass({ $0! == value }))
+        }
+      }
+      
+      context("when calling onCancel") {
+        beforeEach {
+          for idx in 0..<cancelSentinels.count {
+            request.onCancel {
+              cancelSentinels[idx] = true
+            }
+          }
+        }
+        
+        it("should not call the closures") {
+          expect(cancelSentinels.filter({ $0 == nil}).count).to(equal(cancelSentinels.count))
+        }
+      }
+      
+      context("when calling onFailure") {
+        beforeEach {
+          for idx in 0..<successSentinels.count {
+            request.onFailure { error in
+              failureSentinels[idx] = error
+            }
+          }
+        }
+        
+        it("should not call the closures") {
+          expect(failureSentinels.filter({ $0 == nil }).count).to(equal(failureSentinels.count))
+        }
+      }
+      
+      context("when calling onCompletion") {
+        beforeEach {
+          for idx in 0..<successSentinels.count {
+            request.onCompletion { result in
+              switch result {
+              case .Success(let value):
+                successSentinels[idx] = value
+              case .Error(let error):
+                failureSentinels[idx] = error
+              default:
+                break
+              }
+            }
+          }
+        }
+        
+        it("should not call the closures passing an error") {
+          expect(failureSentinels.filter({ $0 == nil }).count).to(equal(failureSentinels.count))
+        }
+        
+        it("should call the closures passing a value") {
+          expect(successSentinels).to(allPass({ $0! == value }))
+        }
+      }
+    }
+    
     sharedExamples("a Future") { (sharedExampleContext: SharedExampleContext) in
       var future: Future<String>!
       var promise: Promise<String>!
@@ -201,16 +384,76 @@ class FutureTests: QuickSpec {
       var future: Future<String>!
       var associatedPromise: Promise<String>!
       
-      beforeEach {
-        associatedPromise = Promise<String>()
-        future = associatedPromise.future
+      context("when obtained through a promise") {
+        beforeEach {
+          associatedPromise = Promise<String>()
+          future = associatedPromise.future
+        }
+        
+        itBehavesLike("a Future") {
+          [
+            FutureSharedExamplesContext.Future: future,
+            FutureSharedExamplesContext.Promise: associatedPromise
+          ]
+        }
       }
       
-      itBehavesLike("a Future") {
-        [
-          FutureSharedExamplesContext.Future: future,
-          FutureSharedExamplesContext.Promise: associatedPromise
-        ]
+      context("when initialized with a value") {
+        let value = "this is a sync success value"
+
+        beforeEach {
+          future = Future(value)
+        }
+
+        itBehavesLike("success case") {
+          [
+            FutureSharedExamplesContext.Future: future,
+            FutureSharedExamplesContext.Value: value
+          ]
+        }
+      }
+
+      context("when initialized with an optional value and an error") {
+        context("when the optional value is nil") {
+          beforeEach {
+            future = Future(value: nil, error: TestError.SimpleError)
+          }
+
+          itBehavesLike("failure case") {
+            [
+              FutureSharedExamplesContext.Future: future
+            ]
+          }
+        }
+
+        context("when the optional value is not nil") {
+          let value = "this is a sync success value"
+
+          beforeEach {
+            future = Future(value: value, error: TestError.SimpleError)
+          }
+
+          itBehavesLike("success case") {
+            [
+              FutureSharedExamplesContext.Future: future,
+              FutureSharedExamplesContext.Value: value
+            ]
+          }
+        }
+      }
+
+      context("when initialized with an error") {
+        let error = TestError.SimpleError
+        
+        beforeEach {
+          future = Future<String>(error)
+        }
+      
+        itBehavesLike("failure case") {
+          [
+            FutureSharedExamplesContext.Future: future
+          ]
+        }
       }
     }
   }